requirements.txt
fastapi
uvicorn
python-multipart
requests
beautifulsoup4
twilio
sqlmodel
python-dotenv
